# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateTeamRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
    type: object
...
  models.MatchResponse:
    properties:
      awayScore:
        type: integer
      awayTeam:
        type: string
      awayTeamId:
        type: integer
      city:
        type: string
      country:
        type: string
      date:
        type: string
      homeScore:
        type: integer
      homeTeam:
        type: string
      homeTeamId:
        type: integer
      id:
        type: integer
      links:
        items:
          $ref: '#/definitions/models.Link'
        type: array
      neutral:
        type: boolean
      tournament:
        type: string
      tournamentId:
        type: integer
    type: object
  models.MatchesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.MatchResponse'
        type: array
      links:
        items:
          $ref: '#/definitions/models.Link'
        type: array
    type: object
...
  models.TeamResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      links:
        items:
          $ref: '#/definitions/models.Link'
        type: array
      name:
        type: string
    type: object
  models.TeamsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.TeamResponse'
        type: array
      links:
        items:
          $ref: '#/definitions/models.Link'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API for football match and team data with HATEOAS and JWT authentication
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Football API
  version: "1.0"
paths:
...
  /football/matches:
    get:
      description: Get all matches with pagination support
      parameters:
      - default: 50
        description: Number of results per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of matches
          schema:
            $ref: '#/definitions/models.MatchesResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List all matches
      tags:
      - matches
...
  /football/teams:
    get:
      description: Get all national teams with HATEOAS links
      produces:
      - application/json
      responses:
        "200":
          description: List of teams
          schema:
            $ref: '#/definitions/models.TeamsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List all teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create a new national team (requires authentication)
      parameters:
      - description: Team details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Team created
          schema:
            $ref: '#/definitions/models.TeamResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Team already exists
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new team
      tags:
      - teams
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"